refactor: 반복되는 함수 제거 및 UILabel 함수 확장 구현
- 반복되는 메서드 제거
- 다양한 글꼴 폰트 적용을 위한 UILabel 메서드 확장 구현
