+++ GSC-CountDown/GSC-CountDown/Scene/MeasureTaskTimeScene/MeasureTaskTimeViewController.swift
@@ -59,7 +59,7 @@ final class MeasureTaskTimeViewController: UIViewController {
         
         setupLayout()
         setupNavigationController()
-        changeTextWeightSpecificRange(label: taskQuestionLabel, range: "능률의 극대화")
+        UILabel().changeTextWeightSemiboldFontSize16(label: taskQuestionLabel, range: "능률의 극대화")
     }
 }
 
@@ -102,13 +102,5 @@ private extension MeasureTaskTimeViewController {
             $0.trailing.equalToSuperview()
             $0.top.equalTo(separator.snp.bottom).offset(32.0)
         }
-    }
-    
-    func changeTextWeightSpecificRange(label: UILabel, range: String) {
-        guard let text = label.text else { return }
-        let attributeString = NSMutableAttributedString(string: text)
-        let font = UIFont.systemFont(ofSize: 16.0, weight: .semibold)
-        attributeString.addAttribute(.font, value: font, range: (text as NSString).range(of: range))
-        label.attributedText = attributeString
     }
 }
